+++ primefaces/src/main/java/org/primefaces/component/commandbutton/CommandButtonBase.java
@@ -63,8 +63,7 @@ public abstract class CommandButtonBase extends HtmlCommandButton implements Aja
         renderDisabledClick,
         ariaLabel,
         ignoreComponentNotFound,
-        disableOnAjax,
-        inlineAjaxStatus
+        disableOnAjax
     }
 
     public CommandButtonBase() {
@@ -293,18 +292,11 @@ public abstract class CommandButtonBase extends HtmlCommandButton implements Aja
     }
 
     public boolean isDisableOnAjax() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.disableOnAjax, false);
+        return (Boolean) getStateHelper().eval(PropertyKeys.disableOnAjax, true);
     }
 
     public void setDisableOnAjax(boolean disableOnAjax) {
         getStateHelper().put(PropertyKeys.disableOnAjax, disableOnAjax);
     }
 
-    public boolean isInlineAjaxStatus() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.inlineAjaxStatus, false);
-    }
-
-    public void setInlineAjaxStatus(boolean inlineAjaxStatus) {
-        getStateHelper().put(PropertyKeys.inlineAjaxStatus, inlineAjaxStatus);
-    }
 }