+++ primefaces/src/main/java/org/primefaces/component/menubutton/MenuButtonBase.java
@@ -219,7 +219,7 @@ public abstract class MenuButtonBase extends AbstractMenu implements Widget {
     }
 
     public boolean isDisableOnAjax() {
-        return (Boolean) getStateHelper().eval(PropertyKeys.disableOnAjax, false);
+        return (Boolean) getStateHelper().eval(PropertyKeys.disableOnAjax, true);
     }
 
     public void setDisableOnAjax(boolean disableOnAjax) {