+++ primefaces/src/main/resources/META-INF/resources/primefaces/forms/forms.commandlink.js
@@ -46,7 +46,7 @@ PrimeFaces.widget.CommandLink = PrimeFaces.widget.BaseWidget.extend({
             }
         });
 
-        if (this.cfg.disableOnAjax === true) {
+        if (this.cfg.disableOnAjax !== false) {
             $(document).on('pfAjaxSend.' + this.id, function(e, xhr, settings) {
                 if (PrimeFaces.ajax.Utils.isXhrSource($this, settings)) {
                     $this.jq.toggleClass('ui-state-loading');