+++ primefaces/src/main/resources/META-INF/resources/primefaces/menu/menu.menubutton.js
@@ -206,7 +206,7 @@ PrimeFaces.widget.MenuButton = PrimeFaces.widget.TieredMenu.extend({
             }
         });
 
-        if (this.cfg.disableOnAjax === true) {
+        if (this.cfg.disableOnAjax !== false) {
             $(document).on('pfAjaxSend.' + this.id, function(e, xhr, settings) {
                 if ($this.isXhrSource(settings)) {
                     $this.disable();