+++ docs/12_0_0/gettingstarted/whatsnew.md
@@ -9,10 +9,8 @@ This page contains a list of big features. Please check the GitHub issues for al
 * AccordionPanel: added support for `options` and `actions` facets of contained `Tab` components.
 * AccordionPanel: now `ui-state-error` class is added to tabs that contain invalid inputs.
 * AutoComplete: added `lazyModel` attribute for fetching autocomplete-suggestions (instead of completeMethod).
-* CommandButton:
-  * added `disableOnAjax` attribute to disable the button during Ajax requests triggered by it.
-  * added `inlineAjaxStatus` attribute to show a loading icon in the button during Ajax requests triggered by it.
 * BreadCrumb: added `seo` attribute for creating JSON Link Data for better SEO.
+* CommandButton: added `disableOnAjax` attribute to disable the button during Ajax requests triggered by it.
 * CommandLink
   * Now has a widget and can be enabled and disabled by its widget.
   * Added `disableOnAjax` attribute to disable the link during Ajax requests triggered by it.